changed how boundary conditions work in the compressible code
bc_lo and bc_hi are no longer used at all

instead you set bc_vel_lo/hi, bc_mass_lo/hi, and bc_therm_lo/hi

as noted in the inputs files:

  ! NOTE: setting bc_vel to periodic sets all the other bc's to periodic)
  ! bc_vel:   -1 = periodic
  !            1 = slip wall
  !            2 = no-slip wall
  ! bc_mass:  -1 = periodic
  !            1 = wall
  !            2 = reservoir (set bc_Yk or bc_Xk in compressible namelist)
  ! bc_therm: -1 = periodic
  !            1 = adiabatic
  !            2 = isothermal (set with t_lo/hi in common namelist)

+++ src_compressible/src_F90/compressible_namelist.F90
@@ -59,12 +59,12 @@ contains
                                                 bc_Yk_in, bc_Xk_in) &
                                                 bind(C, name="initialize_compressible_namespace")
 
-    double precision, intent(inout) :: bc_mass_lo_in(AMREX_SPACEDIM)
-    double precision, intent(inout) :: bc_mass_hi_in(AMREX_SPACEDIM)
-    double precision, intent(inout) :: bc_therm_lo_in(AMREX_SPACEDIM)
-    double precision, intent(inout) :: bc_therm_hi_in(AMREX_SPACEDIM)
-    double precision, intent(inout) :: bc_vel_lo_in(AMREX_SPACEDIM)
-    double precision, intent(inout) :: bc_vel_hi_in(AMREX_SPACEDIM)
+    integer         , intent(inout) :: bc_mass_lo_in(AMREX_SPACEDIM)
+    integer         , intent(inout) :: bc_mass_hi_in(AMREX_SPACEDIM)
+    integer         , intent(inout) :: bc_therm_lo_in(AMREX_SPACEDIM)
+    integer         , intent(inout) :: bc_therm_hi_in(AMREX_SPACEDIM)
+    integer         , intent(inout) :: bc_vel_lo_in(AMREX_SPACEDIM)
+    integer         , intent(inout) :: bc_vel_hi_in(AMREX_SPACEDIM)
     double precision, intent(inout) :: bc_Yk_in(AMREX_SPACEDIM,LOHI,MAX_SPECIES)
     double precision, intent(inout) :: bc_Xk_in(AMREX_SPACEDIM,LOHI,MAX_SPECIES)
